Add files via upload
Scripts for EWAS based on the 2000HIV cohort.
EWASrlm_vac_all.r used a rlm model to perform the EWAS of vaccination effects based on all populations in the 2000HIV discovery cohort. 
EWASrlm_vac_eur.r used a rlm model to perform the EWAS of vaccination effects based on the Europeans in the 2000HIV discovery cohort. 
EWASrlm_cov_all.r used a rlm model to perform the EWAS of covid infection effects based on all populations in the 2000HIV discovery cohort.
EWASrlm_cov_eur.r used a rlm model to perform the EWAS of covid infection effects based on the Europeans in the 2000HIV discovery cohort.
EWASrlm_iso_all.r used a rlm model to perform the EWAS of lockdown effects based on all populations in the 2000HIV discovery cohort.
EWASrlm_iso_eur.r used a rlm model to perform the EWAS of lockdown effects based on the Europeans in the 2000HIV discovery cohort.
EWASrlm_mRNAvac_all.r used a rlm model to perform the EWAS of RNA vaccine effects based on all populations in the 2000HIV discovery cohort.
EWASrlm_mRNAvac_eur.r used a rlm model to perform the EWAS of RNA vaccine effects based on the Europeans in the 2000HIV discovery cohort.
EWASrlm_mRNAvacOthers_all.r used a rlm model to perform the EWAS to detect the different effects between RNA vaccines and non-RNA vaccines based on all populations in the 2000HIV discovery cohort.
EWASrlm_mRNAvacOthers_eur.r  used a rlm model to perform the EWAS to detect the different effects between RNA vaccines and non-RNA vaccines based on the Europeans in the 2000HIV discovery cohort.